调试优化

+++ web/myadmin/viewsTypes.py
@@ -10,6 +10,7 @@ def typeadd(request):
 	ob = Types.objects.all()
 	context = {'types':ob}
 	return render(request, 'back/typeadd.html', context)
+
 # 分类插入
 def typeinsert(request):
 	try:
@@ -26,24 +27,40 @@ def typeinsert(request):
 		return HttpResponse('<script>alert("添加成功");location.href="/myadmin/typelist"</script>')
 	except:
 		return HttpResponse('<script>alert("添加失败");location.href="/myadmin/typelist"</script>')
+
 # 分类列表
 def typelist(request):
 	context = {'typepage':Getpage(request)}
 	return render(request, 'back/typelist.html', context)
+
 # 分类删除
 def typedel(request, tid):
 	try:
-		ob = Types.objects.get(id = tid)
-		ob.delete()
+		# 判断当前分类下是否存在子类
+		num = Types.objects.filter(pid=tid).count()
+		if num:
+			# 有子类,不能删除
+			return HttpResponse('<script>alert("当前分类下有子类,不能删除");location.href="/myadmin/typelist"</script>')
+		else:
+			# 如果没有子类,还要判断当前类下是否有商品发布,
+			ob = Types.objects.get(id = tid)
+			g = ob.goods_set.all().count()
+			if g:
+				# 当前分类下有商品,不能删除
+				return HttpResponse('<script>alert("当前分类下有商品,不能删除");location.href="/myadmin/typelist"</script>')
+			else:
+				ob.delete()
 		return HttpResponse('<script>alert("删除成功");location.href="/myadmin/typelist"</script>')
 	except:
-		return HttpResponse('<script>alert("删除失败");location.href="/myadmin/typelist"</script>')
+		return HttpResponse('<script>alert("无法删除");location.href="/myadmin/typelist"</script>')
+
 # 分类编辑
 def typeedit(request, tid):
 	ob = Types.objects.get(id = tid)
 	obs = Types.objects.all()
 	context = {'tinfo':ob, 'types':obs}
 	return render(request,'back/typeedit.html', context)
+
 # 分类更新
 def typeupdate(request):
 	try:
@@ -55,6 +72,7 @@ def typeupdate(request):
 		return HttpResponse('<script>alert("修改成功");location.href="/myadmin/typelist"</script>')
 	except:
 		return HttpResponse('<script>alert("修改失败");location.href="/myadmin/typelist"</script>')
+		
 # 分页搜索函数
 def Getpage(request):
 	v = request.GET.get('keywords','')
